Major rewrite of the javascript.
Now it is much easier to add functionality because the new callApi()
function abstracts away the details of sending the request.
Also, REMOVED THE GLOBAL VARIABLES (!) for good practice. Use the
getUser() function to retrieve information of the logged in user.

Rewritten doLogin(), doRegister, and addContact(). The first two
are tested and work.